modules/locationProducer/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Josh Haines", email = "Josh@JoshHaines.com"},
]
dependencies = [
    "protobuf>=3.19.4",
    "grpcio>=1.43.0",
    "kafka-python>=2.0.2",
    "grpcio-tools>=1.43.0",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
